+++ modular/latticeOps.h
@@ -2,9 +2,9 @@
 #define LATTICEOPS_H
 #include <random>
 
-void emptyCluster(bool***cluster,double &L);
-double siteEnergy(double *** lattice,double &L, int &s1, int &s2, int &s3);
-double sinX(double ***lattice,double &L, int &s1, int &s2, int &s3,double &angle);
+void emptyCluster(bool***cluster,long double &L);
+long double siteEnergy(long double *** lattice,long double &L, int &s1, int &s2, int &s3);
+long double sinX(long double ***lattice,long double &L, int &s1, int &s2, int &s3,long double &angle);
 
-double*** newLattice(double L,bool cold,std::uniform_real_distribution<double> &dist,std::mt19937_64 &eng);
+long double*** newLattice(long double L,bool cold,std::uniform_real_distribution<long double> &dist,std::mt19937_64 &eng);
 #endif
