+++ modular/spider.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-void whr( double L,double N_equil,double N_samp,bool cold, double* Temps,int N_temps){
+void whr( long double L,long double N_equil,long double N_samp,bool cold, long double* Temps,int N_temps){
 	int nRuns = N_temps/2;
 	int index = N_temps/4;
 	for (int i = 0; i< nRuns; ++i){
@@ -16,19 +16,19 @@ void whr( double L,double N_equil,double N_samp,bool cold, double* Temps,int N_t
 		wolffHistRun(L,N_equil,N_samp,cold,Temps,N_temps,Temps[index]);
 	}
 }
-void wr( double L,double N_equil,double N_samp,bool cold, double* Temps,int N_temps){
+void wr( long double L,long double N_equil,long double N_samp,bool cold, long double* Temps,int N_temps){
 	//regular run
-	double Temp;
+	long double Temp;
 	for (int i = 0; i< N_temps; ++i){
 		Temp = Temps[i];
 		wolffRun(L,N_equil,N_samp,cold,Temp);
 	}
 }
-double * getTrange(double start, double end, int N){
-	double dt = (end-start)/((double)N-1.0);
-	double *T = new double[N];
+long double * getTrange(long double start, long double end, int N){
+	long double dt = (end-start)/((long double)N-1.0);
+	long double *T = new long double[N];
 	for (int i =0; i< N; ++i){
-		T[i] = start + (double)i*dt;
+		T[i] = start + (long double)i*dt;
 	}
 	return T;
 }
@@ -42,14 +42,14 @@ int main(int argc, char* argv[]){
 	
 	cout.precision(17);
 
-	double runType = stod(argv[1]);
-	double L= stod(argv[2]);
-	double startT= stod(argv[3]);
-	double endT= stod(argv[4]);
-	double Ntemps= stod(argv[5]);
-	double Neq= stod(argv[6]);
-	double Nsamp= stod(argv[7]);
-	double icold= stod(argv[8]);
+	long double runType = stod(argv[1]);
+	long double L= stod(argv[2]);
+	long double startT= stod(argv[3]);
+	long double endT= stod(argv[4]);
+	long double Ntemps= stod(argv[5]);
+	long double Neq= stod(argv[6]);
+	long double Nsamp= stod(argv[7]);
+	long double icold= stod(argv[8]);
 
 	//set cold start bool
 	bool cold = false;
@@ -57,7 +57,7 @@ int main(int argc, char* argv[]){
 		cold = true;
 	}
 	//generate temperature range
-	double *Trange = getTrange(startT,endT,int(Ntemps));
+	long double *Trange = getTrange(startT,endT,int(Ntemps));
 	//determine runtype and run
 	if (runType == 0.0) {
 		wr(L,Neq,Nsamp,cold,Trange,Ntemps);
