+++ modular/wolff.h
@@ -2,7 +2,7 @@
 #define WOLFF_H
 
 #include <random>
-int growCluster(double ***lattice,bool ***cluster, double &L,double &beta, double& TotXMag,double& TotYMag,double& TotEn,double &TotSinX,std::uniform_real_distribution<double> &dist,std::mt19937_64 &eng);
+int growCluster(long double ***lattice,bool ***cluster, long double &L,long double &beta, long double& TotXMag,long double& TotYMag,long double& TotEn,long double &TotSinX,std::uniform_real_distribution<long double> &dist,std::mt19937_64 &eng);
