+++ modular/wolffHistRun.h
@@ -2,5 +2,5 @@
 #define WOLFFHISTRUN_H
 
 
-void wolffHistRun(double L, double N_equil_sweeps, double N_samples,bool cold,double *Temperatures,int N_temps,double runTemp);
+void wolffHistRun(long double L, long double N_equil_sweeps, long double N_samples,bool cold,long double *Temperatures,int N_temps,long double runTemp);
 #endif
