+++ modular/wolffRun.h
@@ -2,5 +2,5 @@
 #define WOLFFRUN_H
 
 
-void wolffRun(double L, double N_equil_sweeps, double N_samples,bool cold,double Temperature);
+void wolffRun(long double L, long double N_equil_sweeps, long double N_samples,bool cold,long double Temperature);
 #endif
